+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Release Version 0.19
+
+**Updates**
+* Improved handling of StringIDs, when those IDs are urls.
+* Added support for Delete on Collections using $filter, See https://github.com/opengeospatial/sensorthings/issues/44
+  `service.observations().query().filter("phenomenonTime lt 2018-01-01T12:00:00Z").delete();`
+
+
 # Release Version 0.18
 
 **Updates**